Fix an annoying bug

+++ src/utils/Utils/math.h
@@ -5,6 +5,6 @@ namespace utils
     class math
     {
         public:
-            static int calculate_padding(const int& count, const int& max_padding_len=60);
+            static int calculate_padding(const size_t& count, const size_t& max_padding_len=60);
     };
 }
